fix(issue): [Feature]: Unify HTML artifact shell across /gsd export and /gsd brief
