+++ src/resources/extensions/visual-brief/tests/visual-brief.test.ts
@@ -88,6 +88,20 @@ test("builds a provider-neutral prompt with output instructions", () => {
 	assert.ok(prompt.indexOf("Gather evidence before writing the page") < prompt.indexOf("Write the HTML file"));
 });
 
+test("generated prompts require the shared GSD HTML shell", () => {
+	const request = parseVisualBriefArgs("diff");
+	assert.ok(request, "diff request should parse");
+
+	const prompt = buildVisualBriefPrompt(request, { outputDir: "/tmp/visual-brief" });
+
+	assert.match(prompt, /## Required HTML shell/);
+	assert.match(prompt, /<span class="logo">GSD<\/span>/);
+	assert.match(prompt, /<span class="kind-chip">Diff Brief<\/span>/);
+	assert.match(prompt, /{{MAIN_HTML}}/);
+	assert.match(prompt, /{{GENERATED_AT}}/);
+	assert.match(prompt, /Do not add new <style> blocks/);
+});
+
 test("returns first-argument completions for visual brief modes", () => {
 	const completions = getVisualBriefCompletions("di");
